+++ client/src/components/ambient/ChartTabs/index.js
@@ -140,7 +140,7 @@ const ChartTabs = () => {
     if (activeView === "columns") {
       chartBody = (
         <div className={styles.columnsClickable} onClick={cycleActiveView} role="button" tabIndex={0}>
-          <HourlyForecastColumns />
+          <HourlyForecastColumns expanded={maximized} />
         </div>
       );
     } else {